Lots of changes. Result: BE successful in one directoion. Playing with debug messages (adding and removing).
Fixed a bug in REA handling although this was not necessary as Mika will soon release his changes
that will override these. However, the bug caused oopses during deconfiguration of network interfaces,
so I fixed it here.

TCP saves user socket into hash table and uses source address as one key. This source address happened to
be the real IPv6 address. This caused the very first connection to fail, as the TCP/IP packet that
was received contained HITs (and thus hashed to a different entry). The real problem was that during
adding of HIT mapping, we did not set our own HIT.

Modified hipconf a bit to be able to actually delete mappings. Seems to work, although I'm a bit sceptical
about it.