Improved TCP connection establishment with HIP, UDP packets also work. TCP connection establishment is now intercepted. The data is stored in associated HA structure. When the HIP HA is established, the TCP connection establishment is then continued. This results in very quick handshake. Roughly in 1 second we have completed both HIP and TCP handshakes.
Without these modifications, the time required would have been 12 seconds.
This is due to the fact that TCP buffers the SYN packet for retransmission and because IPsec processing is
done only when routing the packet. Linux routes the packet already in TCP, so the buffered TCP packet already
has been through IPsec processing phase. Unfortunately at that time, there is no HIP SA (as HIP handshake is not complete yet), 
so IPsec did nothing.
Later when there is an SA, the packet does not reroute, which also means that no IPsec processing is done.
The result was to send clear text TCP packets until 2-3 timeouts (since the peer did not answer to TCP packets with HITs). 
After the timeouts, TCP retries to route the packet and this time (since HIP SA is created) it does IPsec processing and
the TCP packet is encapsulated and encrypted in ESP-packet, as the peer expects it to be. 

Db.c is also updated so that it can free the TCP data, if and when necessary.

UDP packet handling was missing. Added that. Did simple test, worked (don't ask me how).