HIP module did not compile as built-in. Fixed. HIP module did not build if compiled as a non-module. Fixed.
Also, I moved the hook variable "hip_functions" to the net/xfrm-state.c
file to avoid incorrect dependencies in the future. At some point, we
will have also IPv4 support, and therefore it does not make sense to
keep the hook variable in the ipv6 code. However, we will need xfrm
indepently of whether IPv4 or IPv6 is used.