Receiving of I1 works now, but crashes :) Receiving of I1 works now, but crashes:
debug(input.c:1632@hip_handle_i1): hip_handle_i1
debug(output.c:223@hip_xmit_r1): 
Segmentation fault

The reason for the segfault is that the code currently tries to receive 
the I1 with "recvfrom" system call. It returns the address of the peer, 
but not the local address. However, the local address is needed for the 
packet handling code.

Alternatives on how to fix this:

* Find a socket option for quering the receiving address.
* Rewrite the msg receiving based on recvmsg interface.
* We try to match a local address based the largest prefix of the peer 
  address. This heuristic may not always return the same address as 
  in the I1 packet.