Fixed afew bugs. Fixed: * The SPI value for the responder was incorrect. * Triggering of bex happened even when the state was established.
Now, the SAs seem to be in order, but no data gets through. I think
this occurs because the SAs have not been somehow "sealed", i.e., we
still get the acquire messages for the HITs even though they should be
set up already.

I suggest comparing with the userspace/main branch code.