Added sources for querying source address in the userspace. Added sources for querying source routes in the userspace. I had to make a lot of modifications to the iproute/libnetlink sources because they were using static variables, exists and fprintfs. We have to maintain those sources redundantly because they cannot be used exactly as they are and they are not provided in Redhat/Fedora...
I also moved a lot of functions from "beet.c" to "nlink.c". It took me
more than two hours to get everything even compiling...

I have not tested this yet, but if the source address selection works,
it is still not returned correctly. See the end of
iproute.c:iproute_get() for additional info.

As an alternative to finding and filling the source address when
mapping, we could have delayed it until receiving of R1. However, then
we should have updated the SP (does it work?).