Added Kristian's IPv6 del addr patch. Date: Wed, 21 Dec 2005 14:28:25 +0200 From: Kristian Slavov <kristian.slavov@nomadiclab.com> To: Kernel Netdev Mailing List <netdev@vger.kernel.org> Subject: [PATCH] Fix IPv6 address deletion
Hi,

If you add more than one IPv6 address belonging to the same prefix and 
delete the address that was last added, routing table entry for that 
prefix is also deleted.
Tested on 2.6.14.4

To reproduce:
ip addr add 3ffe::1/64 dev eth0
ip addr add 3ffe::2/64 dev eth0
/* wait DAD */
sleep 1
ip addr del 3ffe::2/64 dev eth0
ip -6 route

(route to 3ffe::/64 should be gone)

In ipv6_del_addr(), if ifa == ifp, we set ifa->if_next to NULL, and 
later 
assign ifap = &ifa->if_next, effectively terminating the for-loop.
This prevents us from checking if there are other addresses using the 
same 
prefix that are valid, and thus resulting in deletion of the prefix.
This applies only if the first entry in idev->addr_list is the address 
to 
be deleted.

Signed-off-by: Kristian Slavov <kristian.slavov@nomadiclab.com>