CLOSE works partially. CLOSE succeeds but the SAs are reset only at the receiver of the first CLOSE.
Somebody had added state handling at receive_i2 and it was done 
redundantly in handle_i2. If state handling should be done in 
receive_i2, it should be done systematically in all receive functions 
and removed from handle functions. This time, just removed the state 
handling from receive_i2 in anticipation of the refactoring of the state 
machine.