All of the IPsec SPs and SAs are now flushed when the hipd starts. The flushing is enabled by default. The reason for this is that people are doing some very experimental features on some branches that may crash the daemon. In this case, the SPs and SAs are left in the kernel and they disturb further base exchanges.
Cleaner way is to fetch a list of all SPs and SAs, and kill only those 
that are based on HITs. I'll leave this as an exercise for the person 
who has too much spare time in the project :)

If you don't want to flush the SAs, give -N option for the daemon.