iperf patch v2 from Dmitriy Kuptsov Date: Thu, 17 Apr 2008 14:03:39 +0300 From: Dmitriy Kuptsov To: iperf-users Cc: Miika Komu Subject: Iperf in UDP mode and multihomed hosts Parts/Attachments:    1 Shown     19 lines  Text    2   OK     387 lines  Text (charset: UTF-8) ----------------------------------------
Hi,

As previously we encountered a problem with iperf client sending udp
data grams towards multihomed iperf server. The issue was as follows:

1. Run a server on multihomed machine
2. Run a client in udp mode while using non default ip address for
server
3. Once the server gets the messages on non default ip address the
routing will be held by kernel and most likely default ip will be chosen
as source address in the outgoing messages. 
4. The previous step causes the problems. The client will discard
packets that were send in (4) since there will be no matching rule for
that source ip address.

The patch that we advice should solve this issue.