Suresh Kumar: Fixing unaligned access. Fixing unaligned access. This patch should also be applicable to hipl in general, but you will have to make a decision its veracity (I am not sure if this is the only way to fix this issue, but as far as I can see, packing structs is one of the best way to fix it (== root cause).
Some explanation:
When you cast unaligned pointer to an aligned type, the gcc takes your word and inlines memcpy. But this will generate unaligned trap, but will work in x86 and other processors where unaligned accesses will be fixed automatically, but not in ARM.

For example, the following code will raise SIGBUS when run in ARM
(in foo(), the types are same, even though we have typecasted it from a packed struct, so gcc inlines the memcpy)

#include <string.h>
 
typedef struct {
  unsigned int a;
  unsigned int b;
  unsigned char c;
} s;

typedef struct {
  unsigned int a;
  unsigned int b;
  unsigned char c;
} __attribute__ ((packed)) ust;


void foo(s *cp)
{
  s dst;

  memcpy(&dst, cp, sizeof(s));

  return 0;
}

int main(int k, char *kk[]) {

  ust tt;
  return foo(&tt);
}


One place where this was causing a problem was hadb.c::hip_hadb_find_byhits(). There is still one more place where this problem is present: nlink.c::xfrm_fill_selector(), because  "struct in6_addr" is packed but not any of the xfrm* structs. But I am not sure if I should modify it, so I have disabled optimization for that lib (libhiptool) for now