Moved -D_GNU_SOURCE to code level
We might not want to use funtions within _GNU_SOURCE all over the
project, but (for now) only for the specific call of strlen. Therefore,
I moved the define from AM_CFLAGS to the specific include.

The rationale behind the issue:
As a developer, you should never define system-defines (recognizable by
the prefix __) such as __USE_GNU. These are set by the system and the
system only. However, you can and should set single _ defines such as
_GNU_SOURCE.

System libraries should include feature.h, which in turn translates the
user defines to system defines specific to this very development system.
The system defines are useful for the library to know which symbols to
export e.g., in case it provides additional features to POSIX. Hence,
user defines are useful only, when you want to use extra functionality
provided by the lib.