First step on HEARTBEAT modularization.
Created a new module HEARTBEAT. The optional functionality of sending 
heartbeats per ICMP socket will be encapsulated in this module. Right
now the sending of heartbeats should work, but the receiving not.

This will be fixed as soon as the socket handling is modularized.

The function set_cloexec_flag was renamed to hip_set_cloexec_flag and is
not static anymore. This function is neede by the heartbeat module.