Fixed and tested a library dependency problem originating from the RVS code.
The bug report from Diego Biurrun:

In lib/core/builder.c we have the following code snippet:

#ifdef CONFIG_HIP_RVS
        HIP_DEBUG("Received HIT is ours or we are RVS\n");
#else
        HIP_IFEL(!hip_hidb_hit_is_our(&hip_common->hitr), -EFAULT, "Receiver HIT is not ours\n");
#endif /* CONFIG_HIP_RVS */

If RVS is disabled, this creates a dependency on the function
hip_hidb_hit_is_our(), which is in hipd/hidb.c.  This cannot be right,
libhipcore should not depend on hipd.

I'm not sure what the right solution is, it might involve moving some
functionality around.  I'd appreciate if somebody could fix this, it
would allow us to enable the PISA configuration for the autobuilder.