Fixed out-of-source build and make distcheck.
Now, out-of-source builds and make distcheck are working.

Added project_info.xml to EXTRA_DIST and the generated header files and
Makefile.modules to the DISTCLEANFILES list.

Removed creation of emtpy Makefile.modules in autogen.sh. Instead add
'-' prefix to the include statement. This causes automake to ignore, if
the file not exists.