fixed major security whole in the firewall
before ALL HIP packets matching a rule were treated with the configured
verdict instead of incorporating signature verification results, etc
into the overall verdict. This means that as long as e.g. the HIT of a
packet matched a rule in the hipfw, it was accepted...even if the HOST_ID
did actually NOT match the HIT or if the signature was plain wrong.

NOTE: NOTIFY messages MUST contain a HIP signature. However, they don't
for HIPL and thus are blocked by the firewall! This needs to be fixed.