Added rudimental check, if a required module was disabled.
During initialization is checked, if a required module was disabled per
command line argument. This check is needed to avoid the deactivation
of a required module per runtime.

For now, the check produces an error every time a required module was
disabled. Even if the dependent module was disabled too.