Remove doc/Doxyfile from the dist tarball manually.
The complete doc/ subdirectory is listed in EXTRA_DIST to work around the
braindead way that autotools uses to construct distribution tarballs, namely
only include what has been explicitly listed.  Since this is error-prone, we
add complete directories to EXTRA_DIST and never have to worry about declaring
new files as dist_THIS and dist_THAT.  Of course this breaks down when there
is a generated file in such a subdirectory, such as doc/Doxyfile.

So in summary, this disgusting hack is necessary to work around the issue we
have with the stupid dist* implementation in automake.  Way to go...