Major overhaul of the privilege separation code.
The code based on libcap was completely removed, because
of the low/obsolescent availability of libcap (OpenWRT 
doesnt have it, Maemo has an ancient version). The
"altsep" code (based on straight syscalls) is much more
"portable" across Linux boxes: it just relies on some
kernel headers, no libraries involved.

Furthermore, privilege separation is no longer a config
option: it does not make any sense. It belongs to the 
_runtime_. Therefore it's now enabled by default and can
be disabled in hipd via the -p command line switch.

This overhaul was approved by Rene.