libhipcore: move and fix hip_print_version.
A function named like this shoud _not_ call exit(3), even
it's documented in the doxy-comment (it was not the case).
Such "flow control" decisions should only be taken by its
caller.

Even more, a function with just 2 statements does not
deserve its own source file, therefore relocate it as 
"static inline" to the corresponding header file.