Solved problems in loading network modules on Fedora.
Unlike the crypto modules, it appears that network modules are not 
visible as built-in modules. Suppressed errors of unavailable network 
modules because at least I couldn't find out whether xfrm_user is 
built-in or not in Fedora.

If you don't like this fix, please test and contribute a better way to 
deal with this. I can provide access to a test server if needed. The -m 
option is not a better way because it ignores loading of modules 
altogether.