Sanitize packet context handling (hip_packet_context)
This elaborates on commit 4908.
hip_packet_context's fields are statically allocated now, save for the
message buffers which are allocated eactly once and then reused. The
context is correctly reset after each packet now.

hip_port_pair_t used to be typedef'd to struct hip_stateless_info.
The only instance of the latter was in hip_packet_context, so it was
purged in favor of the less-misleading struct hip_portpair_t.