Fixed hash table bug(s)
HIP's hash tables base on OpenSSL's lhash which requires two call-back functions: one for hashing table entries and one for comparing them in the case of hash collisions.

The current code consistently implemented the comparison functions by comparing hashes of the entries. This is incorrect as it does not meet the semantics required of the comparison function. The effect is that when two different table entries have the same hash, the hash table implementation is not able to distinguish them. In this case it can, for example, return the wrong entry during a lookup. Since some hash tables store HIs, this problem might be security sensitive.

With this change, all comparison functions are made to compare the hash input, i.e., the table entries themselves, instead of their hashes.