CentOS RPM build works again.
Applied a workaround from Diego to fix RPM building for CentOS. This (or 
something similar) to make the "make bin" target work on CentOS.

While the bug fix is not perfect, I have tested that it works on CentOS 
and Fedora. If you have further improvements, please commit and ask me 
to test.

See also bug id #620327 in launchpad and the following link for more 
information:

http://fedoraproject.org/wiki/Packaging/DistTag

The "rhel", "rhl" and "el" tags do not seem to work with CentOS.