HIPD now tells signaling module in HIPFW to add a connection on completion of BEX.
HIPD uses a new message for this. We could also include it in the add sa request from the hipd.
However, when a BEX has previously been done this message is not sent but we still need to tell hipfw of the new connection.
Consequently, a new message type is necessary anyway.