Make sure that the executables created by configure have executable permission.
Apparently autotools does not figure that out on its own, you have to invoke
the chmod manually for each file.  Notice that I tried avoiding multiple
AC_CONFIG_FILES and chmod invocations, but autotools balked and barfed and
was generally loathsome, so don't waste your time trying to improve this.