Changed version number for packets for interprocess communications.
The packet builder functionality is reused for both network and 
internal-process communications. The packet version number separates 
external and internal packets from each other. This is important also 
because we allow longer packet lengths for internal packets.

Internal packets were typed as HIPv2 packets and the value is now 
reserved really for HIPv2. To avoid future compatibility issues with 
HIPv2, I rolled the value for internal packets to 0x10.