Fixed a problem with sync-all maintainer script
Fixed a problem reported by Diego: A successful ping was a precondition 
for making the ssh connection. Now the "condition" is a successful echo 
command, so the ssh connection is made no matter if the host is up or 
not...

I decided that the initial ping was unnecessary, so I removed it.