replace hit->ip mapping mechanism by system defaults
Instead of parsing /etc/hosts ourselves, we now use getaddrinfo in order
to retrieve the ip for a given hit. This enables standard DNS lookups,
in case hip/hosts specifies a valid hostname for a hit. As a fallback,
getaddrinfo will still parse and use information provided in /etc/hosts.

Furthermore, this commit removes any manual lookups in /etc/hosts, as
HIP-specific information should be maintained in hip/hosts.