Added functions to convert between compressed (on the wire) format and uncompressed (internal) format of host identities.
These two functions are needed to fix bug 612029. 
Internally host identities are handled within a struct that reserves place for 4096 bit rsa keys and 64 byte hostnames.
For transmission on the wire unused space should not be sent thus we need to compress the host identity before sending it as a message parameter on the wire.
This is done by funciton hip_build_param_host_id.

When receiving a message that contains a host_id parameter we need to decompress it to the internal format. This is done my function hip_build_host_id_from_param.