Fix missing configuration header inclusion.
Revision 4832, by some refered as the grand #include cleanup - part VIII, removed the inclusion of "config.h" in debug.h.
debug.h defines the preprocessor CONFIG_HIP_DEBUG which is used in debug.h to implement different behaviour with #ifdef.

The the absence of the config.h did not introduce any compilation errors, since the macro syntax was still valid (#ifdef CONFIG_HIP_DEBUG), but introduced an unwanted change in behaviour of the programm. Please pay attention next time if you remove "config.h" and make sure that the there are no macros in use defined in "config.h".