Replace duplicated Debian init scripts by symlinks.
We currently have identical Debian initscripts in both the debian and packaging
subdirectories.  The ones in the Debian directory are copies of the ones in the
packaging subdirectory, but do not share the history with them.  In order to
preserve history the initscripts are moved within the revision control system
to their new location in the debian subdirectory and symlinks to the new
location placed in their stead.