Removed TODO comment for struct representation of ecdsa key rrs.
The representation of key in struct hip_host_id as a union is not worthwile, since it brings only a considerable overhead for rsa and dsa keys (they cannot be represented in a struct since their representation as a key rr changes with their size). To ECDSA the union approach brings only a small advantage when handling ecdsa keys (better handling of the curve id).