fix 8 byte instead of 4 byte seq and ack parameter values
SEQ and ACK are defined to be 4 byte values (+ 4 byte parameter header) in
RFC 5201. Both parameters had an overall length of length of 8 bytes for
the value and 4 bytes for the header (= 12 bytes). The reason was that
hip_calc_param_len() expects to be called with the size of the value and
not the size of the complete TLV. Hence, the TL part has to be substracted
from both hip_seq and hip_ack structs before calling this function.
Otherwise, the 4 byte value will be sent on the wire padded by 4 byte of
memory following the struct.
The issue has been fixed now by substracting the TL length. Furthermore,
all other functions using hip_calc_param_len() have been checked to pass
the correct length on function call.