Fixed function hip_get_dsa_public_key().
This function was missing an actual memcpy (or similar) which would set the contents in the output argument. 
The result from this bug was, that neither the host id algorithm, nor the hostname was copied to the new host id structure. 

This revision fixes the bugs by copying the key_rr header, key_rr data and hostname.
Note: The function hip_get_rsa_public_key(), which implements the same functionality for rsa, can serve as a reference for reviewing this revision.