Debian packaging: Do not clean tree after building Debian packages.
Not cleaning the tree after a build is more convenient because a distclean
is performed, which also removes Makefiles and similar things. The original
reason to clean the tree after a Debian package build was sideeffects on the
autobuilder. This can be worked around by testing package building last there.