Rewrite error handling in hip_create_r1().
Currently error values are set to -1 via HIP_IFEL macros, but in the end
NULL is returned in case of error and the error values are ignored. Change
the HIP_IFEL instances to set the pointer holding the return value to NULL
instead and get rid of the int variable for the error values.

This code will not win a beauty contest, but it eliminates the write-only
integer variable err, which caused warnings with -Wunused-but-set-variable.