revert revision 5724 -> renaming of dummy device
Aparently, the dummy device dummy0 is created automatically on
insertion of the dummy module. The only option to rename the device
actually would be to create another device via udev as described
here: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=430528

An even better way would be to get rid of the dummy device althogther
and to replace it with a tun device. This way we could also get rid of
the Linux-specific raw sockets.