Remove virtual interface address (i.e., the HIT) before assigning it.
Because revision 5723 disabled module unloading, the dummy device was shut down but otherwise left in a non-initialized state.

In particular, the address (HIT) was kept. So on the next start of hipd, assigning the HIT again did not set up the 2001:10::/28 route correctly. Presumably, to netlink it looked like a no-op, because the address is set to the current value after all.
As an added bonus, removing the address probably triggers other side-effects of RTM_NEWADDR that you'd have with a blank interface.