Move processing of hipd messages by hipfw into an own subroutine.
Adding additional code after packet processing was difficult because the original code used early exits, and relocating it to a subroutine was IMHO cleaner than using a label.
This also blends in nicely with the other processing routines now.

The code itself was left mostly intact, sans variable declarations and the addition of return codes.