Split old main() function into entry point and actual firewall code.
As the most useful consequence, main() is now contained in an object file of its own, enabling functionality contained in firewall/firewall.c to be used in unit tests.
So far, linking the respective object resulted in an "symbol main already defined" error, because unit tests are implemented using an alternative entry point.