Move keymat generation code back to its original place.
The keymat must be generated before producing any parameters. This was violated by the previous commit but has been restored now.

Also added some (related) const qualifiers to ensure that the input packet is not modified, avoiding confusion in subsequently called handler functions.
We are striving for modularization in this branch, after all.