Refactor lmod_init_state_items() to avoid code duplication.
This commit introduces lmod_run_functions_on_state(). So far only
lmod_init_state_items() makes use of it but a method that calls
uninitialization functions will soon be added and work in the same way on a
different list. Having a static function which accepts the respective list
as a parameter avoids code duplication.