Add and register function to uninitialize an update_state instance.
hip_update_uninit_state() frees the memory that was allocated in the update_state
instance. Register it as a state_uninit_function so that it gets called on
purging of a host association. Add a note that the name used for registering
the update_state instance has to be the same in both init and uninit function.
This fixes a memory leak as the memory was not free'd before.