Remove HIP_IFEL from esp_tuple_from_esp_info_locator().
gcc 4.6 introduced the new warning -Wunused-but-set-variable which is activated
by default with -Wall. In this case it breaks compilation as the err variable
of the HIP_IFEL is set but never used.
Remove the HIP_IFEL to fix the warning and get rid of an unnecessary call to
free().
Note that this revision does NOT compile with gcc 4.6 as two warnings remain.