Return int instead of bool in insert_connection_from_update().
The err variable of the HIP_IFEL is never used breaking compilation with
gcc 4.6. Change the return value from bool to int to make use of err and
adjust the documentation accordingly. Adjust the check for the return value
in handle_update() as well.
Note that this revision does NOT compile with gcc 4.6 because of one remaining 
warning.