Change return value of update_esp_address() from bool to int.
The err variable of the HIP_IFEL was set but never used breaking compilation
with gcc 4.6. Use it as a return value instead of the bool. Adjust the
documentation and the check for return values of callers handle_i2(), handle_r2(),
and update_esp_tuple().
This fixes the last of the -Wunused-but-set-variable warnings.