Use start-stop-daemon to stop hipd in hipl-daemon.init script.
We already use start-stop-daemon to start hipd and there is good reason to use
it to exit it as well. With the specified --retry 3 option it waits for 3
seconds to allow hipd to terminate gracefully before it simply kills it.

This should fix any issues with package installation where hipd was not yet
properly exited but already started again and failing to do so. Unlike the kill
call start-stop-daemon does actually wait for hipd to terminate.

Move starting and exiting of hipd into functions as well to avoid code duplication.