Add restart action to hipfw openwrt init-script.
By default the openwrt call stops and starts hipfw when 'restart' is called.
As hipfw gets started with the -k option by default it is not exited gracefully.
Define the 'restart' call in the init script and sleep for 2 seconds after
calling killall on hipfw. This should be sufficient to clear the firewall rules
and free any states.