Add sanity check against netlink handle and header in netlink_talk().
netlink_talk() should check if it gets a valid netlink handle and message
header before it tries to access them as it is not guaranteed that they are
properly set up. For example when the initialization fails early and the
default policies are nevertheless attempted to be removed on exit.

This fixes bug 790679 (https://bugs.launchpad.net/hipl/+bug/790679).