Rephrase ambiguous error message in hip_xfrm_policy_delete().
If the call to netlink_talk() fails when a policy is to be deleted
it does not mean that there has been no associated policy. It can
fail for any number of reasons (no socket has been set up, a malformed
message was provided etc.). Thus only state that the policy deletion
failed.