Do not depend on network and local filesystem in debian init script.
To avoid having network communication leaked without being handled by
hipd or hipfw we should not force it to be started after the network
has been established.
We are only working on our local filesystem and should not wait for
the setup of the remote filesystem either.

This commit is a follow up to the review of commit 5952 archived
here: http://www.freelists.org/post/hipl-dev/
      Branch-hiplcorehipltrunk-Rev-5952-Require-network-and-local
      -filesystem-to-be-initialized-in-init-scripts