Do not exit on iptable flush error in hipfw init script.
We are running the scripts with set -e, that means every unchecked
command returning an error exits the script. In this case the
function to flash the firewall rules prematurely exited the script
when the rules had already been flushed.
We now check the return value of the flush_iptables() function and
print a notice if an error occurs.

This commit fixes the error return value of the hipfw in bug 795848.