Don't drop packets with non-local dst HIT if RVS is active
Packets with a non-local destination HIT were being dropped in
hip_receive_control_packet() before the check for whether they should be
relayed.

hip_check_i1() (which would only be called after hip_receive_control_packet())
had redundant code doing the same except with a provision for opportunistic I1
packets.

I removed the redundancy and changed the code to be executed only if we are not
offering RVS or relay service.