Moved update header functionality from midauth.c to helpers.c.
The ability to allow modified packets is part of the functionality of the firewall. In order to allow for modified packets, one needs to update the headers of the packet to the new length.
Thus, the functionality to do so should be part of the firewall and not part of the midauth code, which is only included conditionally.